Add navigation menu and Why page
