+++ src/App.css
@@ -14,7 +14,70 @@
 h1 {
   text-align: center;
   color: #1c1c1e;
-  margin-bottom: 2rem;
+  margin: 0;
+}
+
+.app-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 1rem;
+}
+
+.nav-menu {
+  position: relative;
+}
+
+.hamburger {
+  background: none;
+  border: none;
+  padding: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 3px;
+  cursor: pointer;
+}
+
+.hamburger span {
+  display: block;
+  width: 20px;
+  height: 2px;
+  background: #1c1c1e;
+}
+
+.menu-dropdown {
+  position: absolute;
+  top: 100%;
+  right: 0;
+  background: #fff;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+  border-radius: 6px;
+  padding: 0.5rem 0;
+  z-index: 1000;
+}
+
+.menu-dropdown a {
+  display: block;
+  padding: 0.5rem 1rem;
+  color: #1c1c1e;
+  cursor: pointer;
+  text-decoration: none;
+}
+
+.menu-dropdown a:hover {
+  background: #f0f0f0;
+}
+
+.why-page {
+  max-width: 800px;
+  margin: 0 auto;
+  line-height: 1.6;
+}
+
+.why-page h2 {
+  text-align: center;
+  margin-top: 0;
+  margin-bottom: 1rem;
 }
 
 .main-content {
