+++ src/pages/WhyPage.tsx
@@ -1,0 +1,24 @@
+export default function WhyPage() {
+  return (
+    <div className="why-page">
+      <h2>Why We Built This App</h2>
+      <p>
+        Many online paint visualizers simply overlay a flat color onto your
+        photo. Sites from brands like Benjamin Moore and Dunn-Edwards provide a
+        quick preview, but the results often look unnatural and ignore the way
+        light interacts with real walls.
+      </p>
+      <p>
+        We wanted a more accurate approach to help homeowners truly understand
+        how a color will appear in their own space. By leveraging Meta's Segment
+        Anything Model we can isolate walls with much greater precision and
+        preserve the original lighting and shadows of the room.
+      </p>
+      <p>
+        Our goal is to experiment with modern segmentation techniques and offer
+        a smoother workflow for exploring paint options. We hope this project
+        sparks new ideas for improving the design tools available to everyone.
+      </p>
+    </div>
+  )
+}
